Aufgeräumt und Umstrukturiert
functions.php aufgespalten, in erheblichem Maße umstrukturiert.
