+++ .gitignore
@@ -1,3 +1,3 @@
 
 .DS_Store
-system/dbsettings.php
+system/db_settings.php
