+++ admin/newsblog/contentMenu.php
@@ -9,7 +9,7 @@
 
         <div class="dropdown">
             <button class="btn btn-secondary dropdown-toggle" type="button" id="displayArticlesDropdown" data-bs-toggle="dropdown" aria-expanded="false">
-                <i class="bi-calendar"></i>
+                <i class="bi-card-text"></i>
                 Anzeige
             </button>
             <ul class="dropdown-menu" aria-labelledby="displayArticlesDropdown">
