+++ sitemenu.php
@@ -1,19 +1,22 @@
 
 <?php 
     if($_GET['code']) {
-        include(__DIR__."/system/login/activation.php");
+        include(__DIR__."/system/account/login/activation.php");
     } elseif($_GET['nav'] === 'news') {
+        include(__DIR__."/content/newsblog/newsblog_functions.php");
         showAllNews();
     } elseif($_GET['nav'] === 'info') {
         /* include("info.php"); */
     } elseif($_GET['nav'] === 'profile') {
         include(__DIR__."/system/account/settings.php");
     } elseif($_GET['nav'] === 'calendar') {
-        include(__DIR__."/content/calendar/calendarpage.php");
+        include(__DIR__."/content/calendar/calendar_functions.php");
+        include(__DIR__."/content/calendar/calendar_page.php");
     } elseif($_GET['nav'] === 'gallery') {
         include(__DIR__."/content/gallery/gallery.php");
     } else {
         /* default, vorerst Kalender + Willkommenstext einbinden, später variabel machen */
-        include(__DIR__."/content/eventorganizer/main.php");
+        include(__DIR__."/content/calendar/calendar_functions.php");
+        include(__DIR__."/content/calendar/eventorganizer/widget.php");
     }
 ?>
