--- system/dbsettings.php
@@ -1,9 +1,0 @@
-<?php
-    /* TODO: File has to be written during automatic installation process. 
-    In a real release, there should be no need to manually change the values */
-    $dbpre = "clanms_";
-    $dbhost = "localhost";
-    $dbpw = "IZMDPMrIxp6oT7QnqWKA";
-    $dbuser = "clanms";
-    $db = "clanms";
-?>
